V2X-ATCT/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
astropy=4.3.1=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cmake=3.22.1=h1fce559_0
comm=0.1.4=pypi_0
configargparse=1.7=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.33=py37h6a678d5_0
dash=2.15.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
easydict=1.12=pypi_0
expat=2.6.3=h6a678d5_0
fastjsonschema=2.19.1=pypi_0
fire=0.5.0=pypi_0
flask=2.2.5=pypi_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=h4ff587b_1
google-auth=2.28.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.62.0=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.6=pypi_0
imageio=2.31.2=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=7.34.0=pypi_0
ipywidgets=8.1.2=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_3
jsonschema=4.17.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h568e23c_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcurl=8.2.1=h91b91d3_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.52.0=ha637b67_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
llvmlite=0.32.1=py37hd408876_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.4.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-base=3.3.4=py37h62a2d02_0
matplotlib-inline=0.1.6=pypi_0
meshio=5.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mtest=1.0=pypi_0
mypy=1.4.1=pypi_0
mypy-extensions=1.0.0=pypi_0
natsort=8.4.0=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=2.6.3=pypi_0
numba=0.49.0=py37h0573a6f_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
olefile=0.46=py37_0
open3d=0.17.0=pypi_0
opencv-contrib-python=4.1.0.25=pypi_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=6.1.0=pypi_0
pip=24.0=pypi_0
pkg-about=1.0.8=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.18.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.3=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyc-wheel=1.2.7=pypi_0
pyemd=1.0.0=pypi_0
pyerfa=2.0.0.3=pypi_0
pygalmesh=0.10.7=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyqt5=5.15.10=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.13.0=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.15=haa1d7c7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2024.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.4=pypi_0
rhash=1.4.1=h3c74f83_1
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37h06a4308_0
shapely=2.0.3=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
style=1.1.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.2=pyhd3eb1b0_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.13.1=pypi_0
torchvision=0.2.2.post3=pypi_0
tornado=6.2=py37h5eee18b_0
tqdm=4.66.2=pypi_0
traitlets=5.9.0=pypi_0
trimesh=4.4.1=pypi_0
typed-ast=1.5.5=pypi_0
typing-extensions=4.7.1=pypi_0
update=0.0.1=pypi_0
urllib3=2.0.7=pypi_0
vispy=0.5.3=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.10=pypi_0
xz=5.4.5=h5eee18b_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
z3-solver=4.13.4.0=pypi_0